pyproject.toml
# SPDX-FileCopyrightText: 2025 David Mandelberg <david@mandelberg.org>
#
# SPDX-License-Identifier: Apache-2.0

[build-system]
requires = [
  "setuptools",
]
build-backend = "setuptools.build_meta"

[project]
name = "unimnim"
version = "0"
requires-python = ">=3.12"
dependencies = [
  "jinja2",
  "pyicu",
]

[project.optional-dependencies]
test = [
  "black",
  "isort",
  "mypy",
  "pytest",
  "pytest-cov",
]

[project.scripts]
unimnim = "unimnim.main:main"

[tool.black]
line-length = 80

[tool.coverage.run]
omit = [
  "**/*_test.py",
]

[tool.isort]
profile = "google"
skip_gitignore = true
src_paths = ["."]

[tool.mypy]
disallow_any_generics = true
disallow_untyped_defs = true
packages = ["unimnim"]
warn_redundant_casts = true
warn_unreachable = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
  "icu.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = """\
  --cov=. \
  --cov-branch \
  --cov-report=term-missing \
  """

[tool.setuptools.packages.find]
include = ["unimnim", "unimnim.*"]
